Update pylint and associated config (#6517)
* update pylint from 2.6.5 to 2.13.7
* removed outdated rule disables, added some more
* many small code style changes to match new linter rules

+++ integration_tests/python_modules/dagster-k8s-test-infra/dagster_k8s_test_infra/cluster.py
@@ -118,6 +118,7 @@ def local_port_forward_postgres(namespace):
 
     wait_for_pod(postgres_pod_name, namespace=namespace)
 
+    p = None
     try:
         p = subprocess.Popen(
             [
@@ -161,8 +162,9 @@ def local_port_forward_postgres(namespace):
         yield forward_port
 
     finally:
-        print("Terminating port-forwarding")
-        p.terminate()
+        if p is not None:
+            print("Terminating port-forwarding")
+            p.terminate()
 
 
 @pytest.fixture(scope="session")
